Add cleanup of unused heredocs on parsing
